+++ src/main/java/roomescape/globalException/NotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.global_exception;
+package roomescape.globalException;
 
 import org.springframework.http.HttpStatus;
