+++ src/main/java/roomescape/reservation/repository/ReservationRepositoryImpl.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalException.NotFoundException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation_time.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
