+++ src/main/java/roomescape/reservation_time/repository/ReservationTimeRepositoryImpl.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalException.NotFoundException;
 import roomescape.reservation_time.domain.ReservationTime;
 
 @Repository
