+++ src/main/java/roomescape/reservation_time/service/ReservationTimeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import roomescape.global_exception.BadRequestException;
-import roomescape.global_exception.ConflictException;
+import roomescape.globalException.BadRequestException;
+import roomescape.globalException.ConflictException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation_time.ReservationTimeMapper;
