+++ src/test/java/roomescape/reservation/domain/ReservationTest.java
@@ -7,7 +7,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.global_exception.BadRequestException;
+import roomescape.globalException.BadRequestException;
 import roomescape.reservation.fixture.ReservationFixture;
 import roomescape.reservation_time.domain.ReservationTime;
 import roomescape.reservation_time.fixture.ReservationTimeFixture;
