+++ src/test/java/roomescape/reservation_time/repository/ReservationTimeRepositoryImplTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalException.NotFoundException;
 import roomescape.reservation_time.domain.ReservationTime;
 import roomescape.reservation_time.fixture.ReservationTimeFixture;
