+++ src/test/java/roomescape/reservation_time/service/ReservationTimeServiceTest.java
@@ -14,8 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.global_exception.BadRequestException;
-import roomescape.global_exception.NotFoundException;
+import roomescape.globalException.BadRequestException;
+import roomescape.globalException.NotFoundException;
 import roomescape.reservation.domain.dto.ReservationReqDto;
 import roomescape.reservation.repository.ReservationRepositoryImpl;
 import roomescape.reservation.service.ReservationService;
